Refactor zos_files_for_zowe_sdk to use BaseFilesApi
