+++ src/zos_files/zowe/zos_files_for_zowe_sdk/__init__.py
@@ -11,6 +11,7 @@ Copyright Contributors to the Zowe Project.
 """
 
 from . import constants, exceptions
+from .api import BaseFilesApi
 from .datasets import DatasetOption, Datasets
 from .file_system import FileSystems
 from .files import Files
