+++ src/zos_files/zowe/zos_files_for_zowe_sdk/datasets.py
@@ -15,8 +15,8 @@ from typing import Optional, Any, Union
 
 from requests import Response
 
-from zowe.core_for_zowe_sdk import SdkApi
 from zowe.core_for_zowe_sdk.exceptions import FileNotFound
+from zowe.zos_files_for_zowe_sdk.api import BaseFilesApi
 from zowe.zos_files_for_zowe_sdk.constants import FileType, zos_file_constants
 from zowe.zos_files_for_zowe_sdk.response import DatasetListResponse, MemberListResponse
 
@@ -292,7 +292,7 @@ class DatasetOption:
         return {key.replace("_DatasetOption__", ""): value for key, value in self.__dict__.items() if value is not None}
 
 
-class Datasets(SdkApi):  # type: ignore[misc]
+class Datasets(BaseFilesApi):  # type: ignore[misc]
     """
     Class used to represent the base z/OSMF Datasets API.
 
@@ -307,9 +307,7 @@ class Datasets(SdkApi):  # type: ignore[misc]
     """
 
     def __init__(self, connection: dict[str, Any], log: bool = True) -> None:
-
-        super().__init__(connection, "/zosmf/restfiles/", logger_name=__name__, log=log)
-        self._default_headers["Accept-Encoding"] = "gzip"
+        super().__init__(connection, log=log)
 
     def list(self, name_pattern: str, return_attributes: bool = False) -> DatasetListResponse:
         """
