+++ src/zos_files/zowe/zos_files_for_zowe_sdk/file_system.py
@@ -12,8 +12,8 @@ Copyright Contributors to the Zowe Project.
 
 from typing import Optional, Any
 
-from zowe.core_for_zowe_sdk import SdkApi
 from zowe.zos_files_for_zowe_sdk import constants
+from zowe.zos_files_for_zowe_sdk.api import BaseFilesApi
 
 from .exceptions import InvalidPermsOption, MaxAllocationQuantityExceeded
 from .response import FileSystemListResponse
@@ -21,7 +21,7 @@ from .response import FileSystemListResponse
 _ZOWE_FILES_DEFAULT_ENCODING = constants.zos_file_constants["ZoweFilesDefaultEncoding"]
 
 
-class FileSystems(SdkApi):  # type: ignore
+class FileSystems(BaseFilesApi):  # type: ignore
     """
     Class used to represent the base z/OSMF FileSystems API.
 
@@ -36,8 +36,7 @@ class FileSystems(SdkApi):  # type: ignore
     """
 
     def __init__(self, connection: dict[str, Any], log: bool = True):
-        super().__init__(connection, "/zosmf/restfiles/", logger_name=__name__, log=log)
-        self._default_headers["Accept-Encoding"] = "gzip"
+        super().__init__(connection, log=log)
 
     def create(self, file_system_name: str, options: dict[str, Any] = {}) -> None:
         """
