+++ tests/unit/files/datasets/test_delete.py
@@ -1,7 +1,6 @@
-import re
 from unittest import TestCase, mock
 
-from zowe.zos_files_for_zowe_sdk import Datasets, Files, exceptions
+from zowe.zos_files_for_zowe_sdk import Files
 
 
 class TestDeleteClass(TestCase):
